investment_demo/requirements.txt
snowflake-snowpark-python[pandas]>=1.11.0
snowflake-ml-python>=1.1.0
faker>=20.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
argparse>=1.4.0
